cleanup in routes
