+++ src/hooks/translations.ts
@@ -1,17 +1,5 @@
 import { useFinalLocale } from "./final-locale"
-import { PL } from "../translations/pl"
-import { EN } from "../translations/en"
-import { UA } from "../translations/ua"
-
-export type Locales = "pl-PL" | "uk-UA" | "en-US"
-
-export type Translation = typeof PL
-
-export const translations = {
-  "pl-PL": PL,
-  "uk-UA": UA,
-  "en-US": EN,
-}
+import { translations } from "../translations/definitions"
 
 export const useTranslations = () => {
   const finalLocale = useFinalLocale()
